Split data sources file system tools
